// src/components/molecules/topNav/types.ts
export interface HamburgerProps {
  ariaDetails?: any
  children?: any
  hamburgerMargin?: string
  hamburgerPadding?: string
  hamburgerWidth?: string
  hamburgerHeight?: string
  hamburgerBorder?: string
  hamburgerBorderRadius?: string
  hamburgerTextIndent?: string
  hamburgerPosition?: string
  hamburgerTop?: string
  hamburgerLeft?: string
  hamburgerOpacity?: string
  hamburgerFontSize?: string
  hamburgerColor?: string
  hamburgerBackground?: string
  hamburgerZIndex?: string
  hamburgerBeforePostion?: string
  hamburgerBeforeTop?: string
  hamburgerBeforeRight?: string
  hamburgerBeforeBottom?: string
  hamburgerBeforeLeft?: string
  hamburgerBeforeContent?: string
  hamburgerBeforeFontFamily?: string
  hamburgerBeforeBackground?: string
  hamburgerBeforeColor?: string
}



export interface MenuCloseProps {

}

export interface MenuLinkProps {
  linkColor?: string
  menuLinkHoverColor?: string
  menuLinkFontFamily?: string
}

export interface NavBarProps {
  navBarBoxShadow?: string
  backgroundColor?: string
  navBarBorderBottom?: string
}

export interface SlideMenuProps {
  backgroundColor?: string
  slideMenuBorder?: string
  slideMenuListBorder?: string
  slideMenuListBorderRadius?: string
  slideMenuH2fontFamily?: string
  slideMenuH2color?: string
}

export interface TopNavProps {
  topNavAcolor?: string
  topNavAhoverColor?: string
  topNavAfontFamily?: string
  topNavLinks?: any
  linkColor?: string
  linkHoverColor?: string
  linkFontFamily?: string
  linkFontSize?: string
  linkFontWeight?: string
  linkLetterSpacing?: string
  linkPadding?: string
}